Image-Processing-&-Object-Recognition-Pipeline/requirements.txt
# Web Framework
fastapi==0.123.5
uvicorn==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.1

# ML / Computer Vision
torch==2.9.1
torchvision==0.24.1
ultralytics==8.3.234
sahi==0.11.36
transformers==4.57.3
timm==1.0.22
einops==0.8.1
pillow==12.0.0
pillow-heif==1.3.0
opencv-python==4.11.0.86
numpy==2.3.5


# Utilities
scikit-learn==1.7.2
huggingface-hub==0.36.0
requests==2.32.5

# Google Gemini
google-genai==1.56.0

# Vector Search (FAISS)
faiss-cpu==1.13.0

# Image Enhancement (SwinIR)
basicsr==1.4.2

# Database & Cache
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
redis==5.2.1

# Configuration
pydantic==2.12.5
pydantic-settings==2.12.0
